macos: base class for terminal controller
